tests/wpt/tests/fetch/local-network-access/dedicated-worker.tentative.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>LNA Dedicated Worker tests </title>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/support.sub.js"></script>
<script>
  "use strict";

  promise_test(t => {
    const sourceUrl = resolveUrl("resources/dedicated-worker.html",
        sourceResolveOptions({
            server: Server.HTTPS_PUBLIC,
            permissionName: 'loopback-network',
            permissionValue: 'granted'}));
    sourceUrl.searchParams.append("method", "fetch");

    function checkResult(evt) {
      checkFetchTestResult(evt.data, FetchTestResult.SUCCESS);
      t.done();
    }

    const promise = new Promise((resolve) => {
                      window.addEventListener('message', resolve, {once: true});
                    }).then(checkResult);
    const popup = window.open(sourceUrl);
    t.add_cleanup(() => popup.close());

    return promise;
  }, 'LNA Dedicated Worker Public to Loopback Fetch with permission');

  promise_test(t => {
    const sourceUrl = resolveUrl("resources/dedicated-worker.html",
        sourceResolveOptions({
            server: Server.HTTPS_PUBLIC,
            permissionName: 'loopback-network',
            permissionValue: 'denied'}));
    sourceUrl.searchParams.append("method", "fetch");

    function checkResult(evt) {
      checkFetchTestResult(evt.data, FetchTestResult.FAILURE);
      t.done();
    }

    const promise = new Promise((resolve) => {
                      window.addEventListener('message', resolve, {once: true});
                    }).then(checkResult);
    const popup = window.open(sourceUrl);
    t.add_cleanup(() => popup.close());

    return promise;
  }, 'LNA Dedicated Worker Public to Loopback Fetch without permission');

  promise_test(t => {
    const sourceUrl = resolveUrl("resources/dedicated-worker.html",
        sourceResolveOptions({
            server: Server.HTTPS_PUBLIC,
            permissionName: 'loopback-network',
            permissionValue: 'granted'}));
    sourceUrl.searchParams.append("method", "webtransport");

    function checkResult(evt) {
      assert_equals(evt.data, WebTransportTestResult.SUCCESS);
      t.done();
    }

    const promise = new Promise((resolve) => {
                      window.addEventListener('message', resolve, {once: true});
                    }).then(checkResult);
    const popup = window.open(sourceUrl);
    t.add_cleanup(() => popup.close());

    return promise;
  }, 'LNA Dedicated Worker Public to Loopback WebTransport with permission');

  promise_test(t => {
    const sourceUrl = resolveUrl("resources/dedicated-worker.html",
        sourceResolveOptions({
            server: Server.HTTPS_PUBLIC,
            permissionName: 'loopback-network',
            permissionValue: 'denied'}));
    sourceUrl.searchParams.append("method", "webtransport");

    function checkResult(evt) {
      assert_equals(evt.data, WebTransportTestResult.FAILURE);
      t.done();
    }

    const promise = new Promise((resolve) => {
                      window.addEventListener('message', resolve, {once: true});
                    }).then(checkResult);
    const popup = window.open(sourceUrl);
    t.add_cleanup(() => popup.close());

    return promise;
  }, 'LNA Dedicated Worker Public to Loopback WebTransport without permission');

  promise_test(t => {
    const sourceUrl = resolveUrl("resources/dedicated-worker.html",
        sourceResolveOptions({
            server: Server.HTTPS_PUBLIC,
            permissionName: 'loopback-network',
            permissionValue: 'granted'}));
    sourceUrl.searchParams.append("method", "websocket");

    function checkResult(evt) {
      assert_equals(evt.data, WebsocketTestResult.SUCCESS);
      t.done();
    }

    const promise = new Promise((resolve) => {
                      window.addEventListener('message', resolve, {once: true});
                    }).then(checkResult);
    const popup = window.open(sourceUrl);
    t.add_cleanup(() => popup.close());

    return promise;
  }, 'LNA Dedicated Worker Public to Loopback WebSocket with permission');

  promise_test(t => {
    const sourceUrl = resolveUrl("resources/dedicated-worker.html",
        sourceResolveOptions({
            server: Server.HTTPS_PUBLIC,
            permissionName: 'loopback-network',
            permissionValue: 'denied'}));
    sourceUrl.searchParams.append("method", "websocket");

    function checkResult(evt) {
      assert_equals(evt.data, WebsocketTestResult.FAILURE);
      t.done();
    }

    const promise = new Promise((resolve) => {
                      window.addEventListener('message', resolve, {once: true});
                    }).then(checkResult);
    const popup = window.open(sourceUrl);
    t.add_cleanup(() => popup.close());

    return promise;
  }, 'LNA Dedicated Worker Public to Loopback WebSocket without permission');

</script>
</body>
